requirements.txt
-i https://pypi.org/simple
anora==0.1.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
decorator==4.4.0
defusedxml==0.6.0
django-allauth==0.39.1
django-appconf==1.0.3
django-compressor==2.2
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.1.0
django-gravatar2==1.4.2
django-libsass==0.7
django-markupfield==1.5.1
django-typogrify==1.3.3
django==2.2
djangorestframework==3.9.3
docutils==0.14
fabric==2.4.0
gunicorn==19.9.0
idna==2.8
invoke==1.2.0
libsass==0.18.0
markdown==3.1
more-itertools==7.0.0 ; python_version > '2.7'
networkx==2.3
nltk==3.4.1
oauthlib==3.0.1
paramiko==2.4.2
pluggy==0.9.0
psycopg2==2.8.2
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
pymongo==2.7.2
pynacl==1.3.0
pytest==4.4.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
rcssmin==1.0.6
requests-oauthlib==1.2.0
requests==2.21.0
rjsmin==1.0.12
six==1.12.0
smartypants==2.0.1
snowballstemmer==1.2.1
sqlparse==0.3.0
textblob==0.15.3
unidecode==1.0.23
urllib3==1.24.2
